Add new option showDonebutton.
This adds a new option called showDoneButton. It defaults to YES, so no
change in behaviour for old code using TOWebViewController.

If you set this property to NO, it will not add a done bar button item
when presented modally. So it's up to the user where to place a button to
dismiss the view controller.